LayoutTests/platform/gtk/accessibility/spinbutton-value-expected.txt


42
This tests that a spin button supports range value attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS endsWith(axSpin.intValue, '5') is true
PASS endsWith(axSpin.minValue, '1') is true
PASS endsWith(axSpin.maxValue, '9') is true
FAIL endsWith(axSpin.valueDescription, '5 of 9') should be true. Was false.
PASS axSpin.title is axUntitled.title
PASS successfullyParsed is true

TEST COMPLETE
